Ref api service and  Add unit test
